DRY panel error handling
